+++ pkg/list/linked_list_test.go
@@ -14,7 +14,7 @@
 	limitations under the License.
 */
 
-package pool
+package list
 
 import (
 	"testing"
